Revert "String enum Arguments"
This reverts commit 93785b59e789e45280b0d6c42dadf52c1dd7b5a9.
